+++ release_notes.md
@@ -6,3 +6,5 @@
 - Add JitTrace Files for v4.1046
 - Update Java Worker Version to [2.19.4](https://github.com/Azure/azure-functions-java-worker/releases/tag/2.19.4)
 - Tagging cold start impacted requests (#11542)
+- Add area & error code to script & web host health checks (#11552)
+- Log siteUpdateId during initialization to track rolling update progress (#11527)